Update AZ-301T04_Lab_Mod02_Building Azure IaaS-Based Server Applications by using ARM.md
Updated for various UI changes, and updated a few az commands to include the --output parameter

Modified Exercise 3 to have the cloud shell use an existing resource group then added an instruction later on to create the required resource group for deployment.  The original instructions had the creation step in a part that students were advised to skip.